requirements.txt
asgiref==3.8.1
attrs==25.3.0
boto3==1.37.34
botocore==1.37.34
certifi==2025.1.31
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
jmespath==1.0.1
mysqlclient==2.2.7
outcome==1.3.0.post0
packaging==25.0
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
s3transfer==0.11.4
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
